Replace tab completion generation system.
This commit replaces the `Gitsh::Completer` with a more sophisticated
system.

For an overview of how the new system works, see
`lib/gitsh/tab_completion/README.md`, which is added by this commit.

The new system has two benefits over the old:

1. It makes use of the `Gitsh::Lexer` to split up the command, instead of
   naively splitting the input on spaces. This means we're able to tab
   complete commands that don't appear at the beginning of the line, e.g.
   when entering `add . && com<tab>` or `:echo $(con<tab>`.

2. It uses a non-deterministic finite automaton to understand the context of
   the completion, instead of just distinguishing between commands and
   arguments. While the NFA in this commit doesn't do any more than the old
   completer did, it lays the ground work for much more sophisticated
   context-aware completions in the future.

---

There's some slightly ugly code in the new `Gitsh::TabCompletion::Context`
class to deal with inputs that the `Gitsh::Lexer` cannot handle.
Specifically, any input that ends with an incomplete variable, e.g. `"$"`,
`"${"`, or `"${x"`.

When lexical analysis fails, we modify the input to make it acceptable,
retry, and then undo the modification in the resulting token stream.

It's important that this is a retry: if we pre-emptively tried to understand
if the input was acceptable without running it through the lexer it would be
likely we'd produce false positives, e.g. a trailing `$` character in an
unclosed hard string (`":echo '$"`) would be hard to distinguish from
pathological input without the benefit of the lexer.