Fix visualisation.
Free moves weren't shown when they were included after the regular
transitions. I didn't work out why, but changing the order seems to fix it.
That said, I'm not 100% sure we're not just dropping some of the regular
transitions now. There are a lot of them!

It's possible we're hitting some upper limit to the amount of edges dot can
render, but there's nothing documented.